+++ tests/unit/runtime/compile/test_compile_wrapper.py
@@ -8,7 +8,7 @@ import torch
 
 import deepspeed
 from deepspeed.accelerator import get_accelerator
-from deepspeed.runtime.utils import required_torch_version
+from deepspeed.utils.torch import required_torch_version
 
 from unit.common import DistributedTest
